add plot2svg w/ custom tweaks for exporting to SVG
The escape response figures require a big rectangle with transparency. I use the annotation('rectangle' ...) command to generate this because it has to span across axes. Transparency, however, isn't supported using the normal print or write to file functionality. It only works if you convert it to a raster plot, and even then its not very good.

So here I am using the plot2svg package from FileExchange. The problem is that it adds a mysterious bounding box to my rectangles. I dove into the code and made af ix to turn it off. See plot2svg.m, the line with the comment referreing to Andy's Edits, roughly around line 1800.

I've also added the fillpage program from FileExchange which automatically fill the page before ploting2svg.